Feat: Implement authorization logic for post
1. src/posts/controllers/posts.controller.ts
- 유저가 request의 header로 보낸 token으로부터 얻은 nickname으로 본인이 작성한 게시글만 수정, 삭제할 수 있도록 post service에 nickname전달

2. src/posts/services/posts.service.ts
- post controller으로부터 전달받은 nickname으로 db에서 조회한 게시글의 nickname과 비교해 일치하지 않는다면 에러를 출력

+++ node_level2-3/src/posts/services/posts.service.ts
@@ -1,5 +1,9 @@
 import { CreatePostDto } from './../dto/posts.request.dto';
-import { HttpException, Injectable } from '@nestjs/common';
+import {
+  HttpException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { PutRequestDto } from '../dto/posts.request.dto';
 import { PostsRepository } from '../posts.repository';
 
@@ -27,11 +31,14 @@ export class PostsService {
     return post.readOnlyData;
   }
 
-  async updateOnePost(id: string, body: PutRequestDto) {
+  async updateOnePost(id: string, body: PutRequestDto, nickname: string) {
     const { title, content } = body;
     const post = await this.postsRepository.existsById(id);
     if (!post) {
       throw new HttpException('존재하지 않는 게시글입니다.', 404);
+    }
+    if (post.nickname !== nickname) {
+      throw new UnauthorizedException('접근 불가능한 기능입니다.');
     }
 
     post.title = title;
@@ -42,10 +49,13 @@ export class PostsService {
     return { message: '게시글을 수정하였습니다.' };
   }
 
-  async deletePost(id: string) {
+  async deletePost(id: string, nickname: string) {
     const post = await this.postsRepository.existsById(id);
     if (!post) {
       throw new HttpException('존재하지 않는 게시글입니다.', 404);
+    }
+    if (post.nickname !== nickname) {
+      throw new UnauthorizedException('접근 불가능한 기능입니다.');
     }
 
     await this.postsRepository.deletePost(id);
